refactor(sifrr-dom): Merge creator and updator for SifrrElement and SimpleElement
