+++ packages/browser/sifrr-dom/src/dom/ref.js
@@ -17,7 +17,7 @@ function collect(element, stateMap = element.stateMap, filter) {
 }
 
 function filterTW(TW, node, filter) {
-  if (filter && filter(node)){
+  if (filter){
     node = TW.nextSibling();
     if (!node) {
       TW.parentNode();
@@ -30,7 +30,7 @@ function filterTW(TW, node, filter) {
 TREE_WALKER.roll = function(n, filter = false) {
   let node = this.currentNode;
   while(--n) {
-    node = filterTW(TREE_WALKER, node, filter);
+    node = filterTW(TREE_WALKER, node, filter && filter(node));
   }
   return node;
 };
@@ -39,14 +39,15 @@ function create(node, fxn, filter = false) {
   let indices = [], ref, idx = 0, i = 0;
   TREE_WALKER.currentNode = node;
   while(node && i < 5000) {
+    const f = filter && filter(node);
     // eslint-disable-next-line no-cond-assign
-    if (ref = fxn(node)) {
+    if (ref = fxn(node, f)) {
       indices.push(new Ref(idx+1, ref));
       idx = 1;
     } else {
       idx++;
     }
-    node = filterTW(TREE_WALKER, node, filter);
+    node = filterTW(TREE_WALKER, node, f);
     i++;
   }
