+++ packages/browser/sifrr-dom/src/dom/updateattribute.js
@@ -1,0 +1,17 @@
+// Attribute related gotchas
+function updateAttribute(element, name, newValue) {
+  const fromValue = element.getAttribute(name);
+  if (fromValue != newValue) {
+    // values are always cast to strings
+    if (!newValue) {
+      if (fromValue) element.removeAttribute(name);
+    } else {
+      element.setAttribute(name, newValue);
+    }
+  }
+
+  // select/input's value doesn't change on changing value attribute
+  if (name == 'value' && (element.nodeName == 'SELECT' || element.nodeName == 'INPUT')) element.value = newValue;
+}
+
+module.exports = updateAttribute;
